migrate shared web functions
