+++ services/app/src/actions/User.tsx
@@ -1,10 +1,11 @@
 import * as Raven from 'raven-js';
 import Redux from 'redux';
+import {handleFetchErrors} from 'shared/requests';
 import {AUTH_SETTINGS} from '../Constants';
 import {CordovaLoginPlugin, getGA, getGapi, getWindow} from '../Globals';
 import {AppState, UserState} from '../reducers/StateTypes';
 import {loggedOutUser} from '../reducers/User';
-import {fetchUserQuests, handleFetchErrors} from './Web';
+import {fetchUserQuests} from './Web';
 
 interface LoadGapiResponse {gapi: any; async: boolean; }
