--- services/quests/src/actions/Web.test.tsx
@@ -1,3 +1,0 @@
-describe('Web action', () => {
-  test('Empty', () => { /* TODO */ });
-});
