--- services/quests/src/actions/Web.tsx
@@ -1,6 +1,0 @@
-export function handleFetchErrors(response: any) {
-  if (!response.ok) {
-    throw Error(response.statusText);
-  }
-  return response;
-}
