+++ shared/requests/index.test.tsx
@@ -1,0 +1,3 @@
+describe('Request helpers', () => {
+  test('TODO', () => { /* TODO */ });
+});
